// src/types/UserDetails.ts
type UserDetails = {
  needed: boolean;
  weight: string | undefined;
  age: string | undefined;
  plan_to_get_pregnant: string | undefined;
  portion: string | undefined;
  image: string | undefined;
  email: string | undefined;
  firstName: string | undefined;
  lastName: string | undefined;
};

export default UserDetails;